feat(system): add a between-items timeout on all our SSE connections
This avoid potential connection leak by closing them if nothing happen inside it.
These is very unlickly to happen but as some follow endpoint may have races, for ex an execution ends between the time we fetch it and start listening to its event, this is better to add this as a safety net.