build: remove codecov for ui #10038
As discussed with Bart it is not helping us, it is often red and the conf is not right, and anyway he believes codecov on frontend is not easy and perfect

from Bart: Nuance: I believe that having code coverage on Frontend is a great tool to see where we should track more, but having it as a metric is very unproductive.

The only thing I care about being tested are:

if a component renders at all
what it looks like
... interactions
None of this is covered by code lines.

We could have a demand for 10% coverage of every patch and that's it.